Complete 'identify' function
Leverages Dorita980 3.1.10's addition of the 'find' command to locate Roomba when lost. When identify is requested, Roomba will play a series of tones to help you find it. N.B.: The Identify characteristic/feature is supported on 3rd Party HomeKit apps like Eve and Controller, but not in Apple's own Home app.